compile fix

+++ regamedll/dlls/player.cpp
@@ -9907,7 +9907,7 @@ void EXT_FUNC CBasePlayer::__API_HOOK(OnSpawnEquip)(bool addDefault, bool equipG
 #ifdef REGAMEDLL_ADD
 	if(!m_bIsVIP)
 	{
-		switch (static_cast<ArmorType>(free_armor.value))
+		switch (static_cast<ArmorType>((int)free_armor.value))
 		{
 		case ARMOR_KEVLAR: GiveNamedItem("item_kevlar"); break;
 		case ARMOR_VESTHELM: GiveNamedItem("item_assaultsuit"); break;
